requirements.txt
pandas~=2.2.2
yfinance~=0.2.40
plotly~=5.22.0
numpy~=2.0.0
streamlit~=1.36.0
fyers_apiv3~=3.1.2
python-dotenv~=1.0.1

SQLAlchemy~=2.0.31
pydantic~=2.8.2
utils~=1.0.2
python-dateutil~=2.9.0.post0